// src/components/chart/datafeed/datafeed-api.d.ts
export interface Bar {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume?: number;
}